+++ HISTORY.md
@@ -4,6 +4,10 @@
    STDERR of the Phantom process is now piped to the parent process STDERR
    and labeled as `stderr` so that you can tell it apart from `stdout`. (Mark Stosberg)
 
+ * When `DEBUG=phantom-render-stream` is set, we now also provide the console.log output
+   from the Phantom process (Davis Ford and Mark Stosberg)
+
+
 1.6.0 / 2015-01-29
 ==================